Use the normalize method
