+++ components/Message/Link/Video/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import styles from './index.module.css';
+import { normalizeVideoUrl } from './utilities';
 
 interface Props {
   src: string;
@@ -10,7 +11,7 @@ export default function Video({ src }: Props) {
     <iframe
       className={styles.iframe}
       id="ytplayer"
-      src={src}
+      src={normalizeVideoUrl(src)}
       frameBorder="0"
     ></iframe>
   );
