+++ components/Message/Link/utilities.spec.ts
@@ -1,4 +1,4 @@
-import { isImage, isUrlValid, isVideo, normalizeVideoUrl } from './utilities';
+import { isImage, isUrlValid, isVideo } from './utilities';
 
 describe('isUrlValid', () => {
   describe('when url is valid', () => {
@@ -47,30 +47,6 @@ describe('isVideo', () => {
     it('returns true', () => {
       expect(isVideo('https://google.com/main.css')).toBe(false);
       expect(isVideo('https://google.com/main.js')).toBe(false);
-    });
-  });
-});
-
-describe('normalizeVideoUrl', () => {
-  describe('when href is a youtube watch url', () => {
-    it('returns a youtube embed url', () => {
-      expect(
-        normalizeVideoUrl('https://www.youtube.com/watch?v=tgbNymZ7vqY')
-      ).toBe('https://www.youtube.com/embed/tgbNymZ7vqY');
-      expect(normalizeVideoUrl('https://youtube.com/watch?v=tgbNymZ7vqY')).toBe(
-        'https://www.youtube.com/embed/tgbNymZ7vqY'
-      );
-    });
-  });
-
-  describe('when href is a youtube embed url', () => {
-    it('returns the same url', () => {
-      expect(
-        normalizeVideoUrl('https://www.youtube.com/embed/tgbNymZ7vqY')
-      ).toBe('https://www.youtube.com/embed/tgbNymZ7vqY');
-      expect(normalizeVideoUrl('https://youtube.com/embed/tgbNymZ7vqY')).toBe(
-        'https://youtube.com/embed/tgbNymZ7vqY'
-      );
     });
   });
 });
